Same resolution policy as RECON1's 3272c8e9f, re-run on the current tips
(origin/main d8831de2f, origin/port/recon 0d73e6b60, merge base b5ecb84b2):

  src/ include/ config/ tools/ notes/ attribution.json .gitignore -> origin/main
  port/                                                           -> ours

This exists only so step 3 can be measured against the tree the landing
merge will actually produce. Nothing here is a landing candidate.

Conflict surface on these tips, which is LARGER than the plan's 161:
  165 modify/delete (10 include/ headers, 155 src/) -- the plan says 7 + 154
   69 content conflicts, of which 44 are src/ or include/ and 27 of those
      carry a real port-side change against the merge base
   38 add/add, of which 36 are src/ or include/ paths both sides created
      independently (none exist at the merge base)

Every one of those 239 src/include paths is a place taking main drops
port-side work, and only the 161 were enumerated before.

+++ src/MulVec3Mat4x3.c
@@ -6,8 +6,8 @@ void MulVec3Mat4x3(Vector3* v, Matrix4x3* m, Vector3* dst)
     int y = v->y;
     int x = v->x;
     int z = v->z;
-    int* py = (int*)(int)(((long long)(int)((char*)dst + 4)));
-    int* pz = (int*)(int)(((long long)(int)((char*)dst + 8)));
+    int* py = (int*)(int)((char*)dst + 4);
+    int* pz = (int*)(int)((char*)dst + 8);
 
     dst->x = (int)(((long long)x * m->m[0] + (long long)y * m->m[3] + (long long)z * m->m[6]) >> 12);
     dst->x += m->m[9];
