+++ src/func_020050dc.c
@@ -5,6 +5,6 @@ int func_020050dc(void *obj)
 {
     func_0200cb58(obj, 0x1e);
     func_0200cce4(obj);
-    *(int *)(((long long)(int)((char *)obj + 0x154))) |= 0x100;
+    *(int *)((char *)obj + 0x154) |= 0x100;
     return 1;
 }
