+++ src/func_0203558c.c
@@ -1,4 +1,4 @@
 void func_0203558c(char *self)
 {
-    *(unsigned int *)(((long long)(int)(self + 0x10))) |= 0x1000;
+    *(unsigned int *)(self + 0x10) |= 0x1000;
 }
