+++ src/func_020356d4.c
@@ -1,4 +1,4 @@
 void func_020356d4(char *self)
 {
-    *(unsigned int *)(((long long)(int)(self + 0x10))) |= 0x40;
+    *(unsigned int *)(self + 0x10) |= 0x40;
 }
