+++ src/func_020647a4.c
@@ -1,6 +1,6 @@
 void func_020647a4(char *self, int value)
 {
-    *(int *)(((long long)(int)(self + 0x10))) += 1;
+    *(int *)(self + 0x10) += 1;
     *(int *)(self + 0xc) = value;
 
     int *arr = *(int **)(self + 0x14);
