+++ src/func_020707cc.c
@@ -1,6 +1,6 @@
 double func_020707cc(double x)
 {
-    int a = (int)((long long)(int)&x);
+    int a = (int)(&x);
     int *p = (int *)((long long)(a + 4));
     *p &= 0x7fffffff;
     return x;
