+++ src/func_ov002_020da95c.c
@@ -1,4 +1,4 @@
-#define ADDR(p) ((int *)(((long long)(int)(p))))
+#define ADDR(p) ((int *)(p))
 
 #pragma opt_propagation off
 int func_ov002_020da95c(char *c)
