+++ src/func_ov006_020fc1f8.c
@@ -19,18 +19,18 @@ int func_ov006_020fc1f8(char* self, int idx)
     *(u8*)(self + idx*0x1c + 0x4000 + 0xeb5) = data_ov006_0212eb34[*p6];
   }
   {
-    u8* base = (u8*)((int)self + idx*0x1c + 0x4000);
+    u8* base = (u8*)(self + idx*0x1c + 0x4000);
     if (base[0xeb4] != 0) {
       base[0xeba] = 0;
       return 0;
     }
   }
   {
-    u8* p3 = (u8*)((int)self + 0x4eba);
+    u8* p3 = (u8*)(self + 0x4eba);
     u8 c3 = p3[idx*0x1c];
     if (c3 >= 2) {
       c3 = 0;
-      *(u8*)((int)self + idx*0x1c + 0x4000 + 0xeb3) = c3;
+      *(u8*)(self + idx*0x1c + 0x4000 + 0xeb3) = c3;
       p3[idx*0x1c] = c3;
     }
     return c3;
