+++ src/func_ov007_020b8d48.c
@@ -72,7 +72,7 @@ void func_ov007_020b8d48(int a, int c)
         u32 s = *src++ & palmask;
         *dst = (*dst & mask) | (s << shift);
         if (rem != 0) {
-            u32 *d2 = (u32 *)(int)(((long long)(int)(dst + 8)));
+            u32 *d2 = (u32 *)(int)(dst + 8);
             *d2 = (*d2 & ~mask) | (s >> rshift);
         }
         dst++;
