+++ src/func_ov089_02131df4.c
@@ -26,8 +26,8 @@ void func_ov089_02131df4(char* c, char* player)
     {
         char* pl = *(char**)(c + 0x110);
         /* force separate base materialization like ROM: r2=c+0xb0, r3=pl+0x5c */
-        int* fl = (int*)(unsigned)((unsigned long long)(unsigned)(c + 0xb0));
-        int* s = (int*)(unsigned)((unsigned long long)(unsigned)(pl + 0x5c));
+        int* fl = (int*)(unsigned)(c + 0xb0);
+        int* s = (int*)(unsigned)(pl + 0x5c);
         int t0 = s[0];
         int ev = 0x1d;
         *(int*)(c + 0x5c) = t0;
